Remove menu styles from central governance
