+++ packages/harpguru-core/src/components/menu/menu.tsx
@@ -1,8 +1,10 @@
 import Animated from 'react-native-reanimated'
+import { StyleSheet } from 'react-native'
 import React from 'react'
 
 import { getMenuStylesAndAnimationVals } from '../../utils'
 import { MenuProps } from '../../types'
+import { getSizes } from '../../styles'
 
 type LocalMenuProps = MenuProps & {
   readonly position: 'TOP' | 'BOTTOM'
@@ -15,13 +17,28 @@ export const Menu = ({
   children,
 }: LocalMenuProps): React.ReactElement => {
   const {
-    styles,
     menuSlideXTranslation,
     menuSlideYTranslation,
     menuScale,
     menuBackgroundColor,
     menuOpacity,
+    labelProtrusion,
   } = getMenuStylesAndAnimationVals(hideMenu, hideLabel, position)
+
+  const { 9: borderRadius } = getSizes()
+  const styles = StyleSheet.create({
+    animated: {
+      ...StyleSheet.absoluteFillObject,
+      zIndex: 10,
+    },
+    overlay: {
+      ...StyleSheet.absoluteFillObject,
+      left: labelProtrusion * -1,
+      flexDirection: 'row',
+      justifyContent: 'flex-start',
+      borderRadius,
+    },
+  })
 
   return (
     <Animated.View
